Add `Debugger`
